[shell-release]release version 3.1.0

+++ web/src/main/java/com/zrlog/web/inteceptor/StaticResourceInterceptor.java
@@ -5,12 +5,13 @@ import com.hibegin.http.server.api.HttpResponse;
 import com.hibegin.http.server.api.Interceptor;
 import com.hibegin.http.server.util.PathUtil;
 import com.hibegin.http.server.web.MethodInterceptor;
-import com.zrlog.blog.web.interceptor.BlogArticleInterceptor;
+import com.zrlog.blog.web.interceptor.BlogPageInterceptor;
 import com.zrlog.blog.web.plugin.RequestInfo;
 import com.zrlog.blog.web.plugin.RequestStatisticsPlugin;
 import com.zrlog.blog.web.util.WebTools;
-import com.zrlog.business.cache.CacheService;
 import com.zrlog.business.util.InstallUtils;
+import com.zrlog.common.Constants;
+import com.zrlog.util.ZrLogUtil;
 
 import java.io.File;
 import java.util.Arrays;
@@ -20,7 +21,8 @@ public class StaticResourceInterceptor implements Interceptor {
 
     public static final List<String> staticResourcePath = Arrays.asList("/assets", "/admin/static", "/admin/vendors", "/install/static");
 
-    private final CacheService cacheService = new CacheService();
+    public StaticResourceInterceptor() {
+    }
 
     @Override
     public boolean doInterceptor(HttpRequest request, HttpResponse response) throws Exception {
@@ -36,15 +38,15 @@ public class StaticResourceInterceptor implements Interceptor {
             //静态文件进行拦截
             if (staticFile.isFile() && staticFile.exists()) {
                 //缓存静态资源文件
-                if (cacheService.isCacheableByRequest(request)) {
+                if (Constants.zrLogConfig.getCacheService().isCacheableByRequest(request)) {
                     response.addHeader("Cache-Control", "max-age=31536000, immutable"); // 1 年的秒数
                 }
                 response.writeFile(staticFile);
                 return false;
             }
-            if (BlogArticleInterceptor.catGeneratorHtml(actionKey)) {
-                File cacheFile = cacheService.loadCacheFile(request);
-                if (cacheFile.exists()) {
+            if (BlogPageInterceptor.catGeneratorHtml(actionKey)) {
+                File cacheFile = Constants.zrLogConfig.getCacheService().loadCacheFile(request);
+                if (cacheFile.exists() && !ZrLogUtil.isStaticBlogPlugin(request)) {
                     response.writeFile(cacheFile);
                     return false;
                 }
